requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
torch==2.1.0
torchaudio==2.1.0
sounddevice==0.4.6
soundfile==0.12.1
numpy==1.24.3
transformers==4.35.0
requests==2.31.0
scipy==1.11.3
google-generativeai==0.3.0